Basic input, movement updates and rendering of sprites implemented

+++ TGA_Engine_Project/Source/Game/source/PacketComponents/InputComponent.hpp
@@ -10,8 +10,8 @@ public:
 
   uint16_t entityIdentifier = 0;
   
-  int8_t xAxis = 0;
-  int8_t yAxis = 0;
+  float xAxis = 0;
+  float yAxis = 0;
 };
 
 inline InputComponent::InputComponent()
